requirements.txt
mcp>=0.1.6
qdrant-client>=1.9.0
google-api-python-client>=2.129.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
docx>=0.2.4
python-docx>=1.1.2
PyPDF2>=3.0.1
httpx>=0.27.0
